Fix accumulating precision errors in delta encoding.
Features with high-precision coordinates and lines or polygons containing many points can accumulate an error due to the delta encoding method, which only considers the previous point when computing the delta:

  delta[n] = x[n] - x[n - 1]

Instead use the sum of all the previous deltas:

  delta[n] = x[n] - sum(delta[0 .. n-1])

This prevents rounding errors from accumulating. A cumulative sum is kept for speed instead of recomputing the sum each time.